requirements.txt
charset-normalizer==3.4.4
colorama==0.4.6
docx2pdf==0.1.8
greenlet==3.3.2
lxml==6.0.2
pillow==12.1.1
playwright==1.58.0
pyee==13.0.1
python-docx==1.2.0
pywin32==311
reportlab==4.4.10
tqdm==4.67.3
typing_extensions==4.15.0
